requirements.txt
# LocalMartAI Project Requirements (Updated for 2025)
# Core Django dependencies
Django==5.0.8
djangorestframework==3.15.2

# Database - Use PyMySQL instead of mysqlclient for better Windows compatibility
PyMySQL==1.1.1

# AI/ML dependencies (Latest stable versions)
torch==2.5.1
transformers==4.46.2
scikit-learn==1.7.1
lightgbm==4.5.0
pandas==2.2.3
numpy==2.1.3
joblib==1.4.2

# Additional utilities
Pillow==11.0.0
python-decouple==3.8

# Development dependencies (optional)
django-cors-headers==4.4.0
requests==2.32.3
